+++ Demo/src/main/java/com/braintreepayments/demo/DemoActivity.java
@@ -48,18 +48,17 @@ public class DemoActivity extends AppCompatActivity implements ActivityCompat.On
     public BraintreeClient getBraintreeClient() {
         // lazily instantiate braintree client in case the demo has been reset
         if (braintreeClient == null) {
+            boolean useAppLink = Settings.getPayPalLinkType(this).equals(getString(R.string.paypal_app_link));
+            Uri appLinkUri = null;
+            if (useAppLink) {
+                appLinkUri = Uri.parse("https://mobile-sdk-demo-site-838cead5d3ab.herokuapp.com/");
+            }
             if (Settings.useTokenizationKey(this)) {
                 String tokenizationKey = Settings.getTokenizationKey(this);
-
-                boolean useAppLink = Settings.getPayPalLinkType(this).equals(getString(R.string.paypal_app_link));
-                Uri appLinkUri = null;
-                if (useAppLink) {
-                    appLinkUri = Uri.parse("https://mobile-sdk-demo-site-838cead5d3ab.herokuapp.com/");
-                }
                 braintreeClient = new BraintreeClient(this, tokenizationKey, null, appLinkUri);
             } else {
                 braintreeClient =
-                    BraintreeClientFactory.createBraintreeClientWithAuthorizationProvider(this);
+                    BraintreeClientFactory.createBraintreeClientWithAuthorizationProvider(this, appLinkUri);
             }
         }
         return braintreeClient;